+++ spyke/graphics/shaderSources/basic.vert
@@ -5,14 +5,14 @@ layout(location=0) in vec3 aPosition;
 
 layout(location=1) in vec4 aColor;
 layout(location=2) in vec2 aTilingFactor;
-layout(location=3) in int aTexIdx;
-layout(location=4) in int aEntId;
+layout(location=3) in float aTexIdx;
+layout(location=4) in float aEntId;
 layout(location=5) in mat4 aTransform;
 
 out vec4 vColor;
 out vec2 vTexCoord;
-out int vTexIdx;
-out int vEntId;
+out float vTexIdx;
+out float vEntId;
 
 layout(std140) uniform uMatrices
 {
